Refactor de subcarpetas de la carpeta Shared/Serialization
