+++ Assets/Shared/Repositories/MatchesRepositoryJson.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using TopicTwister.Shared.DTOs;
 using TopicTwister.Shared.Interfaces;
 using TopicTwister.Shared.Repositories.IdGenerators;
-using TopicTwister.Shared.Serialization;
+using TopicTwister.Shared.Serialization.Shared;
 using UnityEngine;
