+++ Assets/Shared/Serialization/Deserializers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fecf553b67b93a04e9cb12b6f26d5402
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
