+++ Assets/Shared/Serialization/Serializers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee8f0722e2f0a5040bb2c18bb12c480a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
