+++ Assets/Shared/Serialization/Shared.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 35a536bc4efb84e43a4b06d9892fc88a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
